requirements.txt
flet>=0.21.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
fpdf>=1.7.2
openpyxl>=3.1.0
google-generativeai>=0.3.0
cryptography>=41.0.0
